Rebuild Reef Scoring Scene
Changed the reef scoring scene from UI to a different scene. This scene is planned to be loaded additively on top of the original scene.
Added new reef parts with three different status: normal, selected, and warning.
The part is going to appear as "selected" if the coral icon is currently on top of it.
For now, pressing space (Action) will cause the selected reef part to be marked as "warning." This will be changed to be "full" (having a coral already) later.